Update Detail new version
