+++ UI/PhoneShop/src/app/features/auth/services/auth.service.ts
@@ -20,28 +20,28 @@ export class AuthService {
     private cookieService: CookieService
   ) { }
 
-  login(request: LoginRequest): Observable<LoginResponse>{
-    return this.http.post<LoginResponse>(`https://localhost:7026/api/Auth/login`,{
+  login(request: LoginRequest): Observable<LoginResponse> {
+    return this.http.post<LoginResponse>(`https://localhost:7026/api/Auth/login`, {
       email: request.email,
       password: request.password
     });
   }
 
-  setUser(user: User): void{
+  setUser(user: User): void {
     this.$user.next(user);
     localStorage.setItem('user-email', user.email);
     localStorage.setItem('user-roles', user.roles.join(','));
   }
 
-  user(): Observable<User | undefined>{
+  user(): Observable<User | undefined> {
     return this.$user.asObservable();
   }
 
-  getUser(): User | undefined{
+  getUser(): User | undefined {
     const email = localStorage.getItem('user-email');
     const roles = localStorage.getItem('user-roles');
 
-    if(email && roles){
+    if (email && roles) {
       const user: User = {
         email: email,
         roles: roles.split(',')
@@ -51,9 +51,9 @@ export class AuthService {
     return undefined;
   }
 
-  logout(): void{
+  logout(): void {
     localStorage.clear();
-    this.cookieService.delete('Authorization','/');
+    this.cookieService.delete('Authorization', '/');
     this.$user.next(undefined);
   }
