+++ WebAPI/WebAPI/Repository/IRepository/IUserDetailRepository.cs
@@ -1,0 +1,16 @@
+﻿using WebAPI.ViewModel.UserInformation;
+﻿using Microsoft.AspNetCore.Mvc;
+using WebAPI.Models;
+
+namespace WebAPI.Repository.IRepository
+{
+    public interface IUserDetailRepository
+    {
+        Task<UserViewModel> GetUserByEmail(string email);
+
+        Task<AddressViewModel> GetAddressByUserId(Guid id);
+        Task UpdateAddress(Guid userId, AddressViewModel updatedAddress);
+        Task UpdateUser(Guid userId, UserViewModel updatedUser);
+        Task<bool> ChangePassword(string email, string oldPassword, string newPassword);
+    }
+}
